requirements.txt
accelerate==0.30.1
clip_anytorch==2.6.0
diffusers==0.27.2
hpsv2==1.2.0
matplotlib==3.9.0
ml_collections==0.1.1
numpy==1.26.4
opencv_python==4.9.0.80
peft==0.11.1
Pillow==10.3.0
prompts==0.0.1
pyrallis==0.3.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.0
wandb==0.17.0
